Add static proxify method to ObservableEntity. - Clean up specs.
